+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2026-01-25
+
+### Changed
+- Renamed package from `stt-service` to `aeo-ptt`
+- Renamed Python module from `stt_service` to `aeo_ptt`
+- Renamed CLI commands from `stt-server`/`stt-client` to `aeo-ptt-server`/`aeo-ptt-client`
+- Renamed scripts to match new naming convention
+
 ## [0.1.0] - 2026-01-25
 
 ### Added